lots o levels, not all perfectly integrated but pretty good for the demo i'm going to do on sat march 17

+++ src/edu/washington/cs/games/ktuite/pointcraft/tools/Pellet.java
@@ -30,7 +30,7 @@ public class Pellet implements org.json.JSONString {
 	public boolean alive;
 	public boolean visible;
 	public boolean hover;
-	public int[] color = new int[3];
+	public float[] color = new float[3];
 	public boolean constructing;
 	public float birthday;
 	public int id;
